# backend/openapi.yaml
openapi: 3.0.3
info:
  title: TradingAgentsWeb WS-12 AI Settings API
  version: 0.2.0
  description: |
    Authoritative backend contract for the WS-20 reworked WS-12 AI settings
    flow. User API keys are stored by the frontend in browser localStorage and
    are sent to the backend only for one request at a time. The backend stores
    only user provider metadata. System default provider credentials remain
    backend-managed and must not be exposed to ordinary users.
servers:
  - url: /
...
paths:
  /api/user/llm-settings:
    get:
      tags:
        - user-llm-settings
      operationId: listUserLLMSettings
      summary: List the current user's AI provider metadata.
      description: |
        Returns saved provider metadata for the authenticated user. User API keys
        are not stored by the backend and are not represented in this response.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User AI provider metadata summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserLLMSettingsResponse"
        "401":
          $ref: "#/components/responses/AuthRequired"
  /api/user/llm-settings/providers:
    post:
      tags:
        - user-llm-settings
      operationId: createUserLLMProvider
      summary: Create user AI provider metadata.
      description: |
        Creates one provider metadata profile for the authenticated user. The
        request body contains no API key. Browser-local key save/replace/clear
        is a frontend responsibility.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLLMProviderCreate"
      responses:
        "201":
          description: Created provider metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserLLMProvider"
        "400":
          $ref: "#/components/responses/DetailBadRequest"
        "401":
          $ref: "#/components/responses/AuthRequired"
        "409":
          description: Duplicate provider profile or default constraint violation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                duplicateProvider:
                  value:
                    error:
                      code: USER_LLM_PROVIDER_DUPLICATE
                      message: A provider profile with this provider_name already exists for this user.
        "422":
          $ref: "#/components/responses/DetailValidationError"
  /api/user/llm-settings/providers/{provider_id}:
    parameters:
      - $ref: "#/components/parameters/UserProviderId"
    patch:
      tags:
        - user-llm-settings
      operationId: updateUserLLMProvider
      summary: Update user AI provider metadata.
      description: |
        Updates editable metadata for one user provider profile. Omitted fields
        are unchanged. This endpoint must not accept or persist user key
        material. `is_default: true` makes this profile the user's only default
        profile.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLLMProviderUpdate"
      responses:
        "200":
          description: Updated provider metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserLLMProvider"
        "400":
          $ref: "#/components/responses/DetailBadRequest"
        "401":
          $ref: "#/components/responses/AuthRequired"
        "404":
          $ref: "#/components/responses/UserProviderNotFound"
        "409":
          description: Provider state conflicts with the requested update.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                defaultDisabled:
                  value:
                    error:
                      code: USER_LLM_PROVIDER_DISABLED
                      message: A disabled provider profile cannot be set as default.
        "422":
          $ref: "#/components/responses/DetailValidationError"
    delete:
      tags:
        - user-llm-settings
      operationId: deleteUserLLMProvider
      summary: Delete user AI provider metadata.
      description: |
        Deletes one provider metadata profile owned by the authenticated user.
        The frontend must separately clear any browser-local key for the same
        user/provider scope.
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/AuthRequired"
        "404":
          $ref: "#/components/responses/UserProviderNotFound"
  /api/user/llm-settings/providers/{provider_id}/test:
    parameters:
      - $ref: "#/components/parameters/UserProviderId"
    post:
      tags:
        - user-llm-settings
      operationId: testUserLLMProvider
      summary: Test a user AI provider using a request-time key.
      description: |
        Tests connectivity for a saved user provider profile. The frontend sends
        the API key from localStorage or a one-time input in this request
        body. The backend uses it only for the test call, records validation
        status/time, and must not persist the key.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLLMConnectionTestRequest"
      responses:
        "200":
          description: Connection test result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LLMConnectionTestResponse"
        "400":
          $ref: "#/components/responses/DetailBadRequest"
        "401":
          $ref: "#/components/responses/AuthRequired"
        "404":
          $ref: "#/components/responses/UserProviderNotFound"
        "422":
          $ref: "#/components/responses/DetailValidationError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserProviderId:
      name: provider_id
      in: path
      required: true
      description: User-owned provider metadata ID.
      schema:
        type: integer
        minimum: 1
  responses:
    AuthRequired:
      description: Authentication is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            authRequired:
              value:
                error:
                  code: AUTH_REQUIRED
                  message: Authentication is required.
...
    DetailBadRequest:
      description: The request is syntactically valid but violates business rules.
      content:
        application/json:
          schema:
            type: object
            required:
              - detail
            properties:
              detail:
                type: string
          examples:
            catalogProviderMissing:
              value:
                detail: Catalog provider not found.
    DetailValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            type: object
            required:
              - detail
            properties:
              detail:
                type: string
          examples:
            validation:
              value:
                detail: Request validation failed.
    UserProviderNotFound:
      description: The requested user provider metadata does not exist or is not owned by the current user.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            notFound:
              value:
                error:
                  code: USER_LLM_PROVIDER_NOT_FOUND
                  message: Provider metadata not found.
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
          enum:
            - AUTH_REQUIRED
            - FORBIDDEN
            - VALIDATION_ERROR
            - USER_LLM_PROVIDER_NOT_FOUND
            - USER_LLM_PROVIDER_DUPLICATE
            - USER_LLM_PROVIDER_DISABLED
            - LLM_PROVIDER_NOT_FOUND
            - SYSTEM_DEFAULT_PROVIDER_NOT_SET
            - SYSTEM_DEFAULT_PROVIDER_INACTIVE
            - SYSTEM_DEFAULT_PROVIDER_CREDENTIAL_MISSING
            - REQUEST_PROVIDER_KEY_REQUIRED
            - REQUEST_PROVIDER_INVALID
            - INVALID_BASE_URL
            - CONNECTION_TEST_FAILED
            - LLM_CONFIG_UNRESOLVED
        message:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
        request_id:
          type: string
          nullable: true
    ValidationStatus:
      type: string
      nullable: true
      enum:
        - untested
        - ok
        - failed
    ProviderProfileType:
      type: string
      enum:
        - catalog
        - custom
    UserLLMSettingsResponse:
      type: object
      required:
        - providers
        - default_provider_id
        - has_legacy_config
      properties:
        providers:
          type: array
          items:
            $ref: "#/components/schemas/UserLLMProvider"
        default_provider_id:
          type: integer
          nullable: true
          description: The user's default provider metadata ID, if configured.
        has_legacy_config:
          type: boolean
          description: True when old UserConfig.last_* values can drive a frontend migration prompt.
        legacy_config:
          $ref: "#/components/schemas/LegacyLLMConfigSummary"
          nullable: true
    LegacyLLMConfigSummary:
      type: object
      required:
        - available
      properties:
        available:
          type: boolean
        last_llm_provider:
          type: string
          nullable: true
        last_backend_url:
          type: string
          nullable: true
    UserLLMProvider:
      type: object
      required:
        - id
        - provider_name
        - provider_type
        - display_name
        - base_url
        - shallow_model
        - deep_model
        - is_enabled
        - is_default
        - last_validated_at
        - last_validation_status
        - created_at
        - updated_at
      properties:
        id:
          type: integer
        provider_name:
          type: string
          description: Lowercase provider slug used by analysis requests.
          example: openai
        provider_type:
          $ref: "#/components/schemas/ProviderProfileType"
        catalog_provider_id:
          type: integer
          nullable: true
          description: Linked system catalog provider ID when provider_type is catalog.
        display_name:
          type: string
          example: My OpenAI-compatible endpoint
        base_url:
          type: string
          format: uri
          example: https://api.openai.com/v1
        shallow_model:
          type: string
          example: gpt-5.5
        deep_model:
          type: string
          example: gpt-5.5
        is_enabled:
          type: boolean
        is_default:
          type: boolean
        last_validated_at:
          type: string
          format: date-time
          nullable: true
        last_validation_status:
          $ref: "#/components/schemas/ValidationStatus"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserLLMProviderCreate:
      type: object
      required:
        - provider_name
        - provider_type
        - display_name
        - base_url
        - shallow_model
        - deep_model
      properties:
        provider_name:
          type: string
          minLength: 1
          maxLength: 100
          pattern: "^[a-zA-Z0-9_-]+$"
        provider_type:
          $ref: "#/components/schemas/ProviderProfileType"
        catalog_provider_id:
          type: integer
          minimum: 1
          nullable: true
        display_name:
          type: string
          minLength: 1
          maxLength: 200
        base_url:
          type: string
          format: uri
          maxLength: 500
        shallow_model:
          type: string
          minLength: 1
          maxLength: 200
        deep_model:
          type: string
          minLength: 1
          maxLength: 200
        is_enabled:
          type: boolean
          default: true
        is_default:
          type: boolean
          default: false
    UserLLMProviderUpdate:
      type: object
      description: At least one field must be supplied.
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 200
        base_url:
          type: string
          format: uri
          maxLength: 500
        shallow_model:
          type: string
          minLength: 1
          maxLength: 200
        deep_model:
          type: string
          minLength: 1
          maxLength: 200
        is_enabled:
          type: boolean
        is_default:
          type: boolean
          description: True makes this profile the user's only default profile.
    UserLLMConnectionTestRequest:
      type: object
      required:
        - api_key
      properties:
        api_key:
          type: string
          minLength: 1
          maxLength: 1000
          writeOnly: true
          description: Request-time user key supplied by the frontend for this test only.
        base_url:
          type: string
          format: uri
          nullable: true
          description: Test-only base URL override. Omit to use saved metadata.
        model:
          type: string
          nullable: true
          description: Optional model name to use for a lightweight test request.
    LLMConnectionTestResponse:
      type: object
      required:
        - valid
        - message
        - last_validated_at
        - last_validation_status
      properties:
        valid:
          type: boolean
        message:
          type: string
        last_validated_at:
          type: string
          format: date-time
        last_validation_status:
          $ref: "#/components/schemas/ValidationStatus"
        details:
          type: object
          additionalProperties: true
          nullable: true
          description: Sanitized diagnostics without secrets.
...